JWT extended

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from models.item import ItemModel
 
 class Item(Resource):
@@ -30,8 +30,8 @@ class Item(Resource):
 
         data = Item.parser.parse_args()
 
-        item = ItemModel(name, data['price'], data['store_id'])
-        # item = ItemModel(name, **data)
+        #item = ItemModel(name, data['price'], data['store_id'])
+        item = ItemModel(name, **data)
 
         try:
             item.save_to_db()
@@ -40,8 +40,11 @@ class Item(Resource):
 
         return item.json(), 201
 
-
+    @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -54,8 +57,8 @@ class Item(Resource):
 
 
         if item is None:
-            item = ItemModel(name, data['price'],data['store_id'])
-            #item = ItemModel(name,**data)
+            #item = ItemModel(name, data['price'],data['store_id'])
+            item = ItemModel(name,**data)
         else:
             item.price = data['price']
 
@@ -66,6 +69,6 @@ class Item(Resource):
 
 class ItemList(Resource):
     def get(self):
-        return {"items" : [item.json() for item in ItemModel.query.all()]}
+    #   return {"items" : [item.json() for item in ItemModel.query.all()]}
     #    return {"items": list(map(lambda x: x.json(), ItemModel.query.all()))}
-    #    return {"items": [x.json() for x in ItemModel.query.all()]}
+        return {"items": [x.json() for x in ItemModel.find_all()]}
